Fix connectivity issues on Android
The INTERNET permission was missing on Android, which was
causing problems.  I also updated the hxudp extension
template, to make it hopefully more friendly with the
latest version of lime.
