--- hxudp/.gitignore
@@ -1,6 +1,0 @@
-bin
-project/obj
-project/all_objs
-project/vc100.pdb
-
-.DS_Store
