--- hxudp/.paths
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<ProjectPaths>
-    <OutputFolder>bin</OutputFolder>
-    <SourceFolder>src</SourceFolder>
-    <SourceFolder>test</SourceFolder>
-</ProjectPaths>
