--- hxudp/.project
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<projectDescription>
-	<name>hxudp</name>
-	<comment></comment>
-	<projects>
-	</projects>
-	<buildSpec>
-	</buildSpec>
-	<natures>
-		<nature>eclihx.core.haxenature</nature>
-	</natures>
-</projectDescription>
