--- hxudp/.settings/eclihx.core.prefs
@@ -1,2 +1,0 @@
-contentassist/buildfile=compile.hxml
-eclipse.preferences.version=1
