--- hxudp/README.md
@@ -1,28 +1,0 @@
-# hxudp
-
-UDP socket for [Haxe](http://haxe.org/).
-
-## Support target
-
-* C++:
-  * Windows (32/64bit)
-  * Mac (32/64bit)
-  * Linux (32/64bit)
-  * Android ([NME](http://www.haxenme.org/))
-  * iOS (armv6/armv7)
-
-## Note
-
-The API is not stable and will be revised. If you want to use it, make a copy into your project folder.
-
-## License
-
-hxudp is distributed under the MIT License, same as openFrameworks which hxudp borrowed code from.
-
-> Copyright (c) 2012 Andy Li
-
-> Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
-
-> The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
-
-> THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
