--- hxudp/compile.hxml
@@ -1,6 +1,0 @@
--cpp bin
--main UdpTest
--cp src
--cp test
-
--cmd (cd bin && ./UdpTest)
