+++ hxudp/haxelib.json
@@ -1,0 +1,13 @@
+{
+  "name": "hxudp",
+  "url": "",
+  "license": "MIT",
+  "tags": [],
+  "description": "",
+  "version": "1.0.0",
+  "releasenote": "",
+  "contributors": [""],
+  "dependencies": {
+
+  }
+}
