+++ project.xml
@@ -5,6 +5,8 @@
 	<app main="Main" path="Export" file="Rime" />
 
 	<android permission="android.permission.WRITE_EXTERNAL_STORAGE"/>
+	<android permission="android.permission.CHANGE_WIFI_MULTICAST_STATE"/>
+	<android permission="android.permission.INTERNET"/>
 	<window fps="60"/>
 	<window orientation="portrait" />
 	<source path="Source" />
